Move API definitions to separate Go module:
This allows CAPT to import the definitions without
needing all the other dependencies of Tinkerbell.

Remove the controller-runtime dependency from
the API module. Coupled controller-runtime
deps are challenging to upgrade across repos.
This will require CAPT, and any other consumers,
to copy how pkg/api builds and registers schemas.
This follows the Go proverb,
"A little copying is better than a little dependency."
