+++ Makefile
@@ -164,7 +164,7 @@ manifests: $(CONTROLLER_GEN_FQP) ## Generate WebhookConfiguration and CustomReso
 
 .PHONY: generate
 generate: $(CONTROLLER_GEN_FQP) ## Generate code containing DeepCopy, DeepCopyInto, and DeepCopyObject method implementations.
-	$(CONTROLLER_GEN_FQP) object:headerFile="pkg/api/boilerplate.go.txt" paths="./..."
+	$(CONTROLLER_GEN_FQP) object:headerFile="script/boilerplate.go.txt" paths="./..."
 	$(MAKE) fmt
 
 .PHONY: dep-graph
