+++ api/v1alpha1/tinkerbell/groupversion_info.go
@@ -6,7 +6,6 @@ package tinkerbell
 
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"sigs.k8s.io/controller-runtime/pkg/scheme"
 )
 
 var (
@@ -14,8 +13,8 @@ var (
 	GroupVersion = schema.GroupVersion{Group: "tinkerbell.org", Version: "v1alpha1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme.
-	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
+	//SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
 
 	// AddToScheme adds the types in this group-version to the given scheme.
-	AddToScheme = SchemeBuilder.AddToScheme
+	//AddToScheme = SchemeBuilder.AddToScheme
 )
