+++ api/v1alpha1/tinkerbell/workflowruleset.go
@@ -3,7 +3,7 @@ package tinkerbell
 import metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 func init() {
-	SchemeBuilder.Register(&WorkflowRuleSet{}, &WorkflowRuleSetList{})
+	//SchemeBuilder.Register(&WorkflowRuleSet{}, &WorkflowRuleSetList{})
 }
 
 // +kubebuilder:subresource:status
