+++ rufio/internal/controller/task.go
@@ -20,7 +20,7 @@ import (
 
 	bmclib "github.com/bmc-toolbox/bmclib/v2"
 	"github.com/go-logr/logr"
-	"github.com/tinkerbell/tinkerbell/pkg/api/v1alpha1/bmc"
+	"github.com/tinkerbell/tinkerbell/api/v1alpha1/bmc"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
