+++ tink/controller/controller.go
@@ -6,8 +6,7 @@ import (
 	"net/netip"
 
 	"github.com/go-logr/logr"
-	"github.com/tinkerbell/tinkerbell/pkg/api/v1alpha1/bmc"
-	v1alpha1 "github.com/tinkerbell/tinkerbell/pkg/api/v1alpha1/tinkerbell"
+	"github.com/tinkerbell/tinkerbell/pkg/api"
 	"github.com/tinkerbell/tinkerbell/tink/controller/internal/workflow"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -22,8 +21,8 @@ import (
 
 var schemeBuilder = runtime.NewSchemeBuilder(
 	clientgoscheme.AddToScheme,
-	v1alpha1.AddToScheme,
-	bmc.AddToScheme,
+	api.AddToSchemeTinkerbell,
+	api.AddToSchemeBMC,
 )
 
 type Config struct {
