+++ tink/controller/internal/workflow/convert.go
@@ -2,7 +2,7 @@ package workflow
 
 import (
 	"github.com/oklog/ulid/v2"
-	v1alpha1 "github.com/tinkerbell/tinkerbell/pkg/api/v1alpha1/tinkerbell"
+	v1alpha1 "github.com/tinkerbell/tinkerbell/api/v1alpha1/tinkerbell"
 	"github.com/tinkerbell/tinkerbell/pkg/proto"
 )
