+++ tink/server/internal/grpc/auto.go
@@ -3,7 +3,7 @@ package grpc
 import (
 	"context"
 
-	"github.com/tinkerbell/tinkerbell/pkg/api/v1alpha1/tinkerbell"
+	"github.com/tinkerbell/tinkerbell/api/v1alpha1/tinkerbell"
 )
 
 type AutoReadCreator interface {
