+++ tink/server/internal/grpc/enrollment_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/cenkalti/backoff/v5"
-	"github.com/tinkerbell/tinkerbell/pkg/api/v1alpha1/tinkerbell"
+	"github.com/tinkerbell/tinkerbell/api/v1alpha1/tinkerbell"
 	"github.com/tinkerbell/tinkerbell/pkg/proto"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
